# speed up the performance of s3select on csv