## Description

speed up the performance of s3select on csv

## Motivation and Context

Currently, the performance of s3select on csv does not scale well, here is our bench result:

we are benching using `select * from s3object` on random 64MB csv file, and benching about 300 seconds, single concurrent and 10 concurrents respectively:

|   |  total requests  |  latency (max/min/avg)  |  
|---|---|---|
|  single concurrent  |  55  |  5717#4859#5541  |   
|  10 concurrents  |  180  |  17758#16792#17395  |   

the fiend behind this is that during csv processing, for each record, we have a lot of slice allocation, this drags the performance, and causes not scale up when benching concurrents 10 times than before.

Here is the bench result after this change:

|   |  total requests  |  latency (max/min/avg)  |  
|---|---|---|
|  single concurrent  |  151|  2032#1959#1995  |   
|  10 concurrents  |  990|  3317#2820#3040  |   


## How to test this PR?

basic cases related with s3select have been tested using our internal tool. It makes no difference.

## Types of changes
- [ ] Bug fix (non-breaking change which fixes an issue)
- [ ] New feature (non-breaking change which adds functionality)
- [ ] Breaking change (fix or feature that would cause existing functionality to change)

## Checklist:
- [ ] Fixes a regression (If yes, please add `commit-id` or `PR #` here)
- [ ] Documentation needed
- [ ] Unit tests needed
- [ ] Functional tests needed (If yes, add [mint](https://github.com/minio/mint) PR # here: )