@yaozongyou 2019-07-19
@harshavardhana I have run mint tests against this change, with the MINT_MODE=full and MINT_MODE=core, all tests run successfully.

as for bench testing, I have changed the original select_benchmark_test.go in this pull request to allow Parallel testing.
here is the result in my environment, the bench command is `go test -cpu 10 -benchtime 10x -bench .`

without this change:
```
goos: linux
goarch: amd64
pkg: github.com/minio/minio/pkg/s3select
BenchmarkSelectAll_100K-10                    10         148708616 ns/op        490568216 B/op   1024938 allocs/op
BenchmarkSelectAll_1M-10                      10        1072269673 ns/op        5022156970 B/op 10489830 allocs/op
BenchmarkSelectAll_2M-10                      10        2092445471 ns/op        10045228640 B/op        20979188 allocs/op
BenchmarkSelectAll_10M-10                     10        10721823363 ns/op       50238088362 B/op        104893983 allocs/op
BenchmarkAggregateCount_100K-10               10          27817792 ns/op        42861648 B/op     512848 allocs/op
BenchmarkAggregateCount_1M-10                 10         192783268 ns/op        436492531 B/op   5243772 allocs/op
BenchmarkAggregateCount_2M-10                 10         380384930 ns/op        872699526 B/op  10486647 allocs/op
BenchmarkAggregateCount_10M-10                10        1804155788 ns/op        4370132792 B/op 52429689 allocs/op
PASS
ok      github.com/minio/minio/pkg/s3select     271.507s
```

after this change:
```
goos: linux
goarch: amd64
pkg: github.com/minio/minio/pkg/s3select
BenchmarkSelectAll_100K-10                    10          49208752 ns/op        26921756 B/op     410598 allocs/op
BenchmarkSelectAll_1M-10                      10         411716247 ns/op        274222429 B/op   4198397 allocs/op
BenchmarkSelectAll_2M-10                      10         795807090 ns/op        549349091 B/op   8396206 allocs/op
BenchmarkSelectAll_10M-10                     10        3871556110 ns/op        2758639835 B/op 41978800 allocs/op
BenchmarkAggregateCount_100K-10               10          16307089 ns/op        16647038 B/op     308043 allocs/op
BenchmarkAggregateCount_1M-10                 10         114014101 ns/op        168045510 B/op   3146588 allocs/op
BenchmarkAggregateCount_2M-10                 10         221642064 ns/op        335827644 B/op   6292333 allocs/op
BenchmarkAggregateCount_10M-10                10        1079064368 ns/op        1685777573 B/op 31458160 allocs/op
PASS
ok      github.com/minio/minio/pkg/s3select     146.305s
```